blog/requirements.txt
asgiref==3.5.0
astroid==2.11.2
colorama==0.4.4
dill==0.3.4
Django==4.0.2
django-froala-editor==4.0.10
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
pylint==2.13.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
wrapt==1.14.0
